refactor(dh): separate backend contracts from factory protocol
Move backend TypedDicts, BacklogBackend, and BacklogConfig into a neutral
backend_types module. Concrete backends no longer depend on backend_protocol,
so the factory uses top-level imports without lazy-import suppressions.
