+++ plugins/development-harness/backlog_core/backends/memory_backend.py
@@ -8,7 +8,8 @@ network access or individual gh_client mocking.
 Usage::
 
     from backlog_core.backends.memory_backend import InMemoryBackend
-    from backlog_core.backend_protocol import BacklogBackend, BacklogConfig, set_config
+    from backlog_core.backend_protocol import set_config
+    from backlog_core.backend_types import BacklogBackend, BacklogConfig
 
     backend = InMemoryBackend()
     set_config(BacklogConfig(backend=backend))
@@ -29,7 +30,7 @@ if TYPE_CHECKING:
     from backlog_core.models import Output, SamTask
 
 from backlog_core import rendering as _rendering
-from backlog_core.backend_protocol import (
+from backlog_core.backend_types import (
     BacklogBackend,
     IssueCommentNode,
     IssueNode,
