+++ plugins/development-harness/sam_schema/core/backends/github_task.py
@@ -25,7 +25,7 @@ from sam_schema.core.task_backend_types import DocumentData, DocumentHandle, Pla
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
-    from backlog_core.backend_protocol import BacklogBackend, IssueNode, LabelNode
+    from backlog_core.backend_types import BacklogBackend, IssueNode, LabelNode
     from github.Repository import Repository
 
 __all__ = ["DocumentBackend", "GitHubTaskProvider"]
